refactor(all): Use existing filename-convention (kebab)
