+++ src/components/mobile/survey-questions.tsx
@@ -21,7 +21,7 @@ import {
     FormMessage,
 } from "~/components/ui/form";
 
-import { idToMoreInfo, idToTextMap } from "~/utils/optionMapping";
+import { idToMoreInfo, idToTextMap } from "~/utils/option-mapping";
 
 import { RadioGroup, RadioGroupItem } from "~/components/ui/radio-group";
